requirements.txt
# Basin Signal Engine - Dependencies
streamlit>=1.28.0
openai>=1.0.0
google-generativeai>=0.3.0
groq>=0.4.0
ollama>=0.1.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pydantic>=2.0.0
audio-recorder-streamlit>=0.0.8
